requirements.txt
aiogram==3.1.1
SQLAlchemy==2.0.22
python-dotenv==1.0.0